shiny/requirements.txt
python-dotenv
psycopg2-binary
sqlalchemy
shiny
numpy
pandas
plotly
plotnine
matplotlib
shinywidgets